+delete classroom btn, -join/createClassroomViews, +Modal
TODO: separate join/create classroom modals into components

-join/createClassroomViews +Modal instead
